support updating of output models refactor model backup code for sharing
git-svn-id: svn://localhost/trunk/ecodata@549 63d11371-6d71-872f-44d1-802e8bf8a75e

+++ grails-app/services/au/org/ala/ecodata/MetadataService.groovy
@@ -19,11 +19,18 @@ class MetadataService {
         return activitiesModel().outputs.find { it.name == name }
     }
 
-    def getDataModel(name) {
-        return cacheService.get(name + '-model',{
-            String filename = (grailsApplication.config.app.external.model.dir as String) + name + '/dataModel.json'
+    def getOutputDataModel(templateName) {
+        return cacheService.get(templateName + '-model',{
+            String filename = (grailsApplication.config.app.external.model.dir as String) + templateName + '/dataModel.json'
             JSON.parse(new File(filename).text)
         })
+    }
+
+    def updateOutputDataModel(model, templateName) {
+        log.debug "updating template name = ${templateName}"
+        writeWithBackup(model, grailsApplication.config.app.external.model.dir, templateName, 'dataModel', 'json')
+        // make sure it gets reloaded
+        cacheService.clear(templateName + '-model')
     }
 
     def getModelName(output, type) {
@@ -45,22 +52,34 @@ class MetadataService {
         })
     }
 
+    def writeWithBackup(content, modelPathRoot, path, filename, extension) {
+        /* build path creating directories as required */
+        // assume root ends with a path separator
+        String filePath = modelPathRoot + path + (path.endsWith('/') ? '' : '/')
+        // get new or existing file
+        def f = new File(filePath + filename + '.' + extension)
+
+        // create dirs as required
+        new File(f.getParentFile().getAbsolutePath()).mkdirs()
+
+        if (f.exists()) {
+            // create a backup of the file appending the current timestamp to the name
+            Date date = new Date();
+            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss");
+            def backupFilename = filePath + filename + "-" + sdf.format(date) + '.' + extension
+            f.renameTo(new File(backupFilename))
+        }
+
+        // write the new content
+        f.withWriter { out ->
+            out.write content as String
+        }
+    }
+
     def updateActivitiesModel(model) {
-        // get the existing model
-        String filename = (grailsApplication.config.app.external.model.dir as String) + 'activities-model.json'
-        def f = new File(filename)
-
-        // create a backup of the file appending the current timestamp to the name
-        Date date = new Date();
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss");
-        def backupFilename = (grailsApplication.config.app.external.model.dir as String) + 'activities-model' +
-                "-" + sdf.format(date) + ".json"
-        f.renameTo(new File(backupFilename))
-
-        // write the new model
-        f.write(model)
-
+        writeWithBackup(model, grailsApplication.config.app.external.model.dir, '', 'activities-model', 'json')
         // make sure it gets reloaded
         cacheService.clear('activities-model')
     }
+
 }
